// core/domain/enums/event.enum.ts
export enum EventTypeEnum {
    SPLIT = "split",
    REVERSE_SPLIT = "reverse_split",
    SPIN_OFF = "spin_off",
    BANKRUPTCY = "bankruptcy",
    RIGHTS_ISSUE = "rights_issue",
    SHARE_BUYBACK = "share_buyback",
    MERGER = "merger",
    ACQUISITION = "acquisition",
    DELISTING = "delisting",
    DIVIDEND = "dividend",
}